[RENAME] #56 - ClubMeetingResponse 파일 위치 변경
- 파일 위치 변경
- ClubMeetingRequest ->  Builder 삭제
