+++ src/main/java/com/core/linkup/club/clubmeeting/request/ClubMeetingRequest.java
@@ -1,10 +1,7 @@
 package com.core.linkup.club.clubmeeting.request;
 
-import lombok.Builder;
-
 import java.time.LocalDateTime;
 
-@Builder
 public record ClubMeetingRequest(
         String title,
         LocalDateTime date,
